[MOVE/#13] Extension 패키지 생성
