+++ app/src/main/java/org/sopt/dosopttemplate/presentation/main/home/list/MyInfoViewHolder.kt
@@ -2,7 +2,7 @@ package org.sopt.dosopttemplate.presentation.main.home.list
 
 import androidx.recyclerview.widget.RecyclerView
 import org.sopt.dosopttemplate.databinding.ItemMyInfoBinding
-import org.sopt.dosopttemplate.util.setOnSingleClickListener
+import org.sopt.dosopttemplate.util.extension.setOnSingleClickListener
 
 class MyInfoViewHolder(
     val binding: ItemMyInfoBinding,
