+++ app/src/main/java/org/sopt/dosopttemplate/util/base/BindingActivity.kt
@@ -7,7 +7,7 @@ import androidx.annotation.LayoutRes
 import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.DataBindingUtil
 import androidx.databinding.ViewDataBinding
-import hideKeyboard
+import org.sopt.dosopttemplate.util.extension.hideKeyboard
 
 abstract class BindingActivity<T : ViewDataBinding>(
     @LayoutRes private val layoutResId: Int,
