+++ app/src/main/java/org/sopt/dosopttemplate/util/extension/ContextExt.kt
@@ -1,3 +1,5 @@
+package org.sopt.dosopttemplate.util.extension
+
 import android.app.Activity
 import android.content.Context
 import android.view.View
