+++ app/src/main/java/org/sopt/dosopttemplate/util/extension/StringExt.kt
@@ -1,3 +1,3 @@
-package org.sopt.dosopttemplate.util
+package org.sopt.dosopttemplate.util.extension
 
 fun String.checkLength(min: Int, max: Int): Boolean = this.length in min..max
